requirements.txt
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-openai==0.1.8
langchain-cohere==0.1.5
langchain-voyageai==0.1.1
streamlit==1.35.0
pymongo==4.7.2
pandas==2.0.3
python-dotenv==1.0.1
argparse==1.4.0